Initial import
Initial implementation of a full-featured, helpful CLI script for IntelliJ IDEA and various similar JetBrains IDEs that use the same command-line format and options (which is many, but by no means all, of JetBrains’ products).

The CLI script is essentially a template, which gets lightly edited by a setup script as it’s copied into place, which appears to be how JetBrains Toolbox works as well.